refactor(wallet): move payment domain under transactions
