+++ client/src/components/pages/Store/Wallet/Transactions/Payment/components/PaymentSuccessContent.jsx
@@ -12,7 +12,7 @@ import { useTranslations } from "next-intl";
 import { useCurrency } from "@/components/hooks/useCurrency";
 import { CopyButton } from "@/components/shared/CopyButton";
 
-import { formatSats } from "../../utils/formatters";
+import { formatSats } from "../../../utils/formatters";
 import { useSatsToFiatEstimate } from "../hooks/useSatsToFiatEstimate";
 import { formatFiat } from "../utils/formatFiat";
