+++ client/src/components/pages/Store/Wallet/Transactions/Payment/components/PendingPaymentContent.jsx
@@ -13,7 +13,7 @@ import { useTranslations } from "next-intl";
 
 import { useCurrency } from "@/components/hooks/useCurrency";
 
-import { formatSats } from "../../utils/formatters";
+import { formatSats } from "../../../utils/formatters";
 import { usePaymentAmountInput } from "../hooks/usePaymentAmountInput";
 import { formatFiat as formatFiatValue } from "../utils/formatFiat";
