+++ client/src/components/pages/Store/Wallet/Transactions/Payment/utils/formatFiat.js
