+++ client/src/components/pages/Store/Wallet/Transactions/Payment/utils/validateBolt11Invoice.js
