+++ client/src/components/pages/Store/Wallet/Transactions/__tests__/SendTab.test.jsx
@@ -5,7 +5,7 @@ import BitcoinPriceService from "@/services/bitcoinPriceService";
 import * as walletService from "@/services/walletService";
 import { I18nProvider } from "@i18n/I18nProvider";
 
-import { PaymentTab } from "../../Payment";
+import { PaymentTab } from "../Payment";
 
 jest.mock("@heroui/react", () => {
   const actual = jest.requireActual("@heroui/react");
